+++ packages/pigeon/example/app/ios/Runner/Messages.g.swift
@@ -406,7 +406,7 @@ class MessageFlutterApi: MessageFlutterApiProtocol {
             let details: String? = nilOrValue(listResponse[2])
             continuation.resume(
               throwing: PigeonError(code: code, message: message, details: details))
-          } else if listResponse[0] == nil {
+          } else if listResponse[0] == nil || listResponse[0] is NSNull {
             continuation.resume(
               throwing: PigeonError(
                 code: "null-error",