+++ packages/pigeon/platform_tests/swift_concurrency_test_plugin/darwin/swift_concurrency_test_plugin/Sources/swift_concurrency_test_plugin/Enum.gen.swift
@@ -360,7 +360,7 @@ class EnumApi2Flutter: EnumApi2FlutterProtocol {
             let message: String? = nilOrValue(listResponse[1])
             let details: String? = nilOrValue(listResponse[2])
             continuation.resume(throwing: EnumError(code: code, message: message, details: details))
-          } else if listResponse[0] == nil {
+          } else if listResponse[0] == nil || listResponse[0] is NSNull {
             continuation.resume(
               throwing: EnumError(
                 code: "null-error",