+++ packages/pigeon/platform_tests/swift_concurrency_test_plugin/darwin/swift_concurrency_test_plugin/Sources/swift_concurrency_test_plugin/Message.gen.swift
@@ -550,7 +550,7 @@ class MessageFlutterSearchApi: MessageFlutterSearchApiProtocol {
             let details: String? = nilOrValue(listResponse[2])
             continuation.resume(
               throwing: MessageError(code: code, message: message, details: details))
-          } else if listResponse[0] == nil {
+          } else if listResponse[0] == nil || listResponse[0] is NSNull {
             continuation.resume(
               throwing: MessageError(
                 code: "null-error",