+++ packages/pigeon/platform_tests/swift_concurrency_test_plugin/darwin/swift_concurrency_test_plugin/Sources/swift_concurrency_test_plugin/MultipleArity.gen.swift
@@ -173,7 +173,7 @@ class MultipleArityFlutterApi: MultipleArityFlutterApiProtocol {
             let details: String? = nilOrValue(listResponse[2])
             continuation.resume(
               throwing: MultipleArityError(code: code, message: message, details: details))
-          } else if listResponse[0] == nil {
+          } else if listResponse[0] == nil || listResponse[0] is NSNull {
             continuation.resume(
               throwing: MultipleArityError(
                 code: "null-error",