+++ packages/pigeon/platform_tests/swift_concurrency_test_plugin/darwin/swift_concurrency_test_plugin/Sources/swift_concurrency_test_plugin/NonNullFields.gen.swift
@@ -458,7 +458,7 @@ class NonNullFieldFlutterApi: NonNullFieldFlutterApiProtocol {
             let details: String? = nilOrValue(listResponse[2])
             continuation.resume(
               throwing: NonNullFieldsError(code: code, message: message, details: details))
-          } else if listResponse[0] == nil {
+          } else if listResponse[0] == nil || listResponse[0] is NSNull {
             continuation.resume(
               throwing: NonNullFieldsError(
                 code: "null-error",