+++ packages/pigeon/platform_tests/swift_concurrency_test_plugin/darwin/swift_concurrency_test_plugin/Sources/swift_concurrency_test_plugin/ProxyApiTests.gen.swift
@@ -3142,7 +3142,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3196,7 +3196,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3250,7 +3250,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3304,7 +3304,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3358,7 +3358,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3412,7 +3412,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3467,7 +3467,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3521,7 +3521,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3578,7 +3578,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3632,7 +3632,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3686,7 +3686,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -4211,7 +4211,7 @@ protocol PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(