+++ packages/pigeon/platform_tests/test_plugin/darwin/test_plugin/Sources/test_plugin/ProxyApiTests.gen.swift
@@ -3124,7 +3124,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3178,7 +3178,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3232,7 +3232,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3286,7 +3286,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3340,7 +3340,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3394,7 +3394,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3449,7 +3449,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3503,7 +3503,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3559,7 +3559,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3613,7 +3613,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -3667,7 +3667,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(
@@ -4192,7 +4192,7 @@ final class PigeonApiProxyApiTestClass: PigeonApiProtocolProxyApiTestClass {
           let message: String? = nilOrValue(listResponse[1])
           let details: String? = nilOrValue(listResponse[2])
           completion(.failure(ProxyApiTestsError(code: code, message: message, details: details)))
-        } else if listResponse[0] == nil {
+        } else if listResponse[0] == nil || listResponse[0] is NSNull {
           completion(
             .failure(
               ProxyApiTestsError(